move everything into simdjson_internal
