+++ cpp/src/arrow/json/CMakeLists.txt
@@ -21,7 +21,6 @@ add_arrow_test(test
                chunker_test.cc
                converter_test.cc
                from_string_test.cc
-               object_parser_test.cc
                parser_test.cc
                reader_test.cc
                PREFIX
