+++ cpp/src/parquet/encryption/key_material.h
@@ -21,13 +21,9 @@
 
 #include "parquet/platform.h"
 
-namespace arrow {
-namespace json {
-namespace internal {
+namespace arrow::internal {
 class ObjectParser;
-}  // namespace internal
-}  // namespace json
-}  // namespace arrow
+}  // namespace arrow::internal
 
 namespace parquet::encryption {
 
@@ -86,8 +82,7 @@ class PARQUET_EXPORT KeyMaterial {
 
   static KeyMaterial Parse(const std::string& key_material_string);
 
-  static KeyMaterial Parse(
-      const ::arrow::json::internal::ObjectParser* key_material_json);
+  static KeyMaterial Parse(const ::arrow::internal::ObjectParser* key_material_json);
 
   /// This method returns a json string that will be stored either inside a parquet file
   /// or in a key material store outside the parquet file.
