+++ cpp/src/parquet/encryption/local_wrap_kms_client.cc
@@ -17,16 +17,15 @@
 
 #include <string_view>
 
-#include "arrow/json/object_parser.h"
-#include "arrow/util/json_writer_internal.h"
 #include "arrow/util/secure_string.h"
+#include "arrow/util/simdjson_internal.h"
 
 #include "parquet/encryption/key_toolkit_internal.h"
 #include "parquet/encryption/local_wrap_kms_client.h"
 #include "parquet/exception.h"
 
-using ::arrow::json::JsonWriter;
-using ::arrow::json::internal::ObjectParser;
+using ::arrow::internal::JsonWriter;
+using ::arrow::internal::ObjectParser;
 using ::arrow::util::SecureString;
 
 namespace parquet::encryption {
